+++ synthesize/state.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
-from collections.abc import Collection, Iterator
+from collections import defaultdict
+from collections.abc import Collection, Iterator, Mapping
 from dataclasses import dataclass
 from enum import Enum
 
@@ -9,8 +10,9 @@ from networkx import DiGraph, ancestors, descendants
 from synthesize.config import After, Flow, FlowNode
 
 
-class FlowNodeStatus(Enum):
+class Status(Enum):
     Pending = "pending"
+    Waiting = "waiting"
     Running = "running"
     Succeeded = "succeeded"
     Failed = "failed"
@@ -20,7 +22,7 @@ class FlowNodeStatus(Enum):
 class FlowState:
     graph: DiGraph
     flow: Flow
-    statuses: dict[str, FlowNodeStatus]
+    statuses: dict[str, Status]
 
     @classmethod
     def from_flow(cls, flow: Flow) -> FlowState:
@@ -35,53 +37,54 @@ class FlowState:
         return FlowState(
             graph=graph,
             flow=flow,
-            statuses={id: FlowNodeStatus.Pending for id in graph.nodes},
+            statuses={id: Status.Pending for id in graph.nodes},
         )
 
-    def running_nodes(self) -> Collection[FlowNode]:
-        return tuple(
-            self.flow.nodes[id]
-            for id, status in self.statuses.items()
-            if status is FlowNodeStatus.Running
-        )
+    def nodes_by_status(self) -> Mapping[Status, Collection[FlowNode]]:
+        d = defaultdict(list)
+        for id, s in self.statuses.items():
+            d[s].append(self.flow.nodes[id])
+        return d
 
     def ready_nodes(self) -> Collection[FlowNode]:
         return tuple(
             self.flow.nodes[id]
             for id in self.graph.nodes
-            if self.statuses[id] is FlowNodeStatus.Pending
-            and all(self.statuses[a] is FlowNodeStatus.Succeeded for a in ancestors(self.graph, id))
+            if self.statuses[id] is Status.Pending
+            and all(
+                self.statuses[a]
+                in (
+                    Status.Succeeded,
+                    Status.Waiting,
+                )
+                for a in ancestors(self.graph, id)
+            )
         )
 
-    def mark_success(self, node: FlowNode) -> None:
-        self.statuses[node.id] = FlowNodeStatus.Succeeded
+    def mark_success(self, *nodes: FlowNode) -> None:
+        self.mark(*nodes, status=Status.Succeeded)
 
-    def mark_failure(self, node: FlowNode) -> None:
-        self.statuses[node.id] = FlowNodeStatus.Failed
+    def mark_failure(self, *nodes: FlowNode) -> None:
+        self.mark(*nodes, status=Status.Failed)
 
-    def mark_pending(self, node: FlowNode) -> None:
-        self.statuses[node.id] = FlowNodeStatus.Pending
+    def mark_pending(self, *nodes: FlowNode) -> None:
+        self.mark(*nodes, status=Status.Pending)
 
-    def mark_descendants_pending(self, node: FlowNode) -> None:
-        for t in _descendants(self.graph, {node.id}):
-            self.statuses[t] = FlowNodeStatus.Pending
+    def mark_running(self, *nodes: FlowNode) -> None:
+        self.mark(*nodes, status=Status.Running)
 
-    def mark_running(self, node: FlowNode) -> None:
-        self.statuses[node.id] = FlowNodeStatus.Running
+    def mark(self, *nodes: FlowNode, status: Status) -> None:
+        for node in nodes:
+            self.statuses[node.id] = status
+
+    def children(self, node: FlowNode) -> Collection[FlowNode]:
+        return tuple(self.flow.nodes[id] for id in self.graph.successors(node.id))
+
+    def descendants(self, node: FlowNode) -> Collection[FlowNode]:
+        return tuple(self.flow.nodes[id] for id in descendants(self.graph, node.id))
 
     def all_done(self) -> bool:
-        return all(status is FlowNodeStatus.Succeeded for status in self.statuses.values())
-
-    def num_nodes(self) -> int:
-        return len(self.graph)
+        return all(status is Status.Succeeded for status in self.statuses.values())
 
     def nodes(self) -> Iterator[FlowNode]:
         yield from self.flow.nodes.values()
-
-
-def _ancestors(graph: DiGraph, nodes: set[str]) -> set[str]:
-    return nodes.union(*(ancestors(graph, n) for n in nodes))
-
-
-def _descendants(graph: DiGraph, nodes: set[str]) -> set[str]:
-    return nodes.union(*(descendants(graph, n) for n in nodes))